musicTribes/requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
colorama==0.4.4
Django==3.1.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.2.0
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap4==2.3.1
django-widget-tweaks==1.4.8
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
pylint==2.6.0
pytz==2020.5
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
